Create separate files for several types
