+++ SuperpositionGame/src/main/scala/superposition/math/DependentKey.scala
@@ -1,0 +1,5 @@
+package superposition.math
+
+trait DependentKey {
+  type Value
+}
