+++ SuperpositionGame/src/main/scala/superposition/math/DependentMap.scala
@@ -1,0 +1,3 @@
+package superposition.math
+
+trait DependentMap[K <: DependentKey] extends DependentMapOps[K, DependentMap[K]]
