+++ SuperpositionGame/src/main/scala/superposition/quantum/StateId.scala
@@ -1,0 +1,7 @@
+package superposition.quantum
+
+import superposition.math.DependentKey
+
+final class StateId[A] extends DependentKey {
+  type Value = A
+}
